Update order callback
Note: The three existing callbacks are now identical apart from the fact that the order function is returned the data named as 'message' whereas the others are sent 'cart'. If this were changed in OrderCart.module.php, all the callbacks could be merged into one. I'm keeping this as is for now though as I may want to add extra feedback for the user.